add kernel cross build support for faster builds on x86_64-linux
This adds the option `raspberry-pi-nix.kernel-build-system`, which can be used to drastically decrease the build times.

Description of the option:

The build system to compile the kernel on.

Only the linux kernel will be cross compiled, while most of the derivations are still pulled from cache.nixos.org.

Use this if you cannot or don't want to use the nix-community cache and either:
  - you are building on an x86_64 system using binfmt_misc for aarch64-linux.
  - or if your x86_64 builder has a better CPU than your aarch64 builder.